📱 Улучшена работа с golang collect coverage
